Move config validation function to separate module
