+++ pkg/cmd/agent/agent.go
@@ -1,30 +1,16 @@
 package agent
 
 import (
-	"errors"
 	"net/url"
 	"strings"
 
 	"github.com/MakeNowJust/heredoc"
 	"github.com/buildkite/cli/v3/internal/config"
 	"github.com/buildkite/cli/v3/pkg/cmd/factory"
+	"github.com/buildkite/cli/v3/pkg/cmd/validation"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
-
-// CheckValidConfiguration returns a function that checks the viper configuration is valid to execute the command
-func CheckValidConfiguration(v *viper.Viper) func(cmd *cobra.Command, args []string) error {
-	var err error
-
-	// ensure the configuration has an API token set
-	if !v.IsSet(config.APITokenConfigKey) {
-		err = errors.New("You must set a valid API token. Run `bk configure`.")
-	}
-
-	return func(cmd *cobra.Command, args []string) error {
-		return err
-	}
-}
 
 func NewCmdAgent(f *factory.Factory) *cobra.Command {
 	cmd := cobra.Command{
@@ -34,7 +20,7 @@ func NewCmdAgent(f *factory.Factory) *cobra.Command {
 		Example: heredoc.Doc(`
 			$ bk agent stop buildkite/018a2b90-ba7f-4220-94ca-4903fa0ba410
 		`),
-		PersistentPreRunE: CheckValidConfiguration(f.Config),
+		PersistentPreRunE: validation.CheckValidConfiguration(f.Config),
 		Annotations: map[string]string{
 			"help:arguments": heredoc.Doc(`
 				An agent can be supplied as an argument in any of the following formats:
